feat: add support for multiple desktops (#151)
* Add support for including multiple desktops

Include windows cloaked by the shell, but exclude "invisible" windows.
This seems to result in pretty much the expected list of windows across
all desktops instead of just the current one.

* Use registry key instead of config setting

Windows already has a setting for the Alt-tab switcher to include
virtual desktops, so we can just use that instead of a config.

* Add configuration for only_current_desktop

Did a bit of cleanup / refactoring as well.

* Combine invisible/visible check and more cleanup

* Move registry check out of loop + minor refactor